refactor: move System Prompt construction and LLM calling from Gateway to Session layer
Gateway no longer directly builds System Prompts or calls LLM Providers.
All LLM interaction is delegated to the new session::llm_caller module.

Changes:
- New session/llm_caller.rs with build_request, call_llm,
  call_llm_streaming, push_messages_with_injection, and SinkUpdater
- session_handler.rs: call_llm delegates to session::llm_caller::call_llm
- session_handler_streaming.rs: delegates to session::llm_caller for
  prompt building and stream opening, removed duplicated SinkUpdater

Source: user
Type: refactor
